Conflicts:
	aplicacoes/gal_npi/src/main/java/br/ufc/npi/gal/model/EstruturaCurricular.java
	aplicacoes/gal_npi/src/main/java/br/ufc/npi/gal/web/ComponenteCurricularController.java
	aplicacoes/gal_npi/src/main/webapp/WEB-INF/views/componente/listar.jsp
	aplicacoes/gal_npi/src/main/webapp/WEB-INF/views/curso/estrutura/adicionar.jsp
	aplicacoes/gal_npi/src/main/webapp/WEB-INF/views/curso/estrutura/editar.jsp

+++ aplicacoes/gal_npi/src/main/java/br/ufc/npi/gal/web/ComponenteCurricularController.java
@@ -3,6 +3,8 @@ package br.ufc.npi.gal.web;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import javax.inject.Inject;
 import javax.servlet.http.HttpSession;
@@ -25,6 +27,8 @@ import br.ufc.npi.gal.model.IntegracaoCurricular;
 import br.ufc.npi.gal.model.Titulo;
 import br.ufc.npi.gal.service.CalculoMetaService;
 import br.ufc.npi.gal.service.ComponenteCurricularService;
+import br.ufc.npi.gal.service.EstruturaCurricularService;
+import br.ufc.npi.gal.service.IntegracaoCurricularService;
 import br.ufc.npi.gal.service.MetaCalculada;
 import br.ufc.npi.gal.service.ResultadoCalculo;
 import br.ufc.npi.gal.service.TituloService;
@@ -36,16 +40,22 @@ public class ComponenteCurricularController {
 
 	@Inject
 	private ComponenteCurricularService componenteCurricularService;
-	
+
 	@Inject
 	private TituloService tituloService;
-	
+
 	@Inject
 	private CalculoMetaService calculoService;
 
 	@Inject
 	private GenericService<Bibliografia> bibliografiaService;
-	
+
+	@Inject
+	private IntegracaoCurricularService integracaoCurricularService;
+
+	private static final String COMPLEMENTAR = "Complementar";
+	private static final String BASICA = "Básica";
+
 	@RequestMapping(value = "/listar")
 	public String listar(ModelMap modelMap, HttpSession session) {
 		modelMap.addAttribute("componentes", this.componenteCurricularService.find(ComponenteCurricular.class));
@@ -66,8 +76,10 @@ public class ComponenteCurricularController {
 	}
 
 	@RequestMapping(value = "/editar", method = RequestMethod.POST)
-	public String atualizar(@Valid ComponenteCurricular componente, BindingResult result,
+	public String atualizar(@ModelAttribute("componente") ComponenteCurricular componente, BindingResult result,
 			RedirectAttributes redirectAttributes) {
+		boolean errors = false;
+
 		if (result.hasErrors()) {
 			return "componente/editar";
 		}
@@ -76,32 +88,40 @@ public class ComponenteCurricularController {
 				componente.getCodigo()) != null) {
 			result.rejectValue("codigo", "Repeat.componente.codigo",
 					"Já existe um componente curricular com esse código");
-			return "componente/editar";
+			errors = true;
 		}
+
 		if (componenteCurricularService.getOutraComponenteCurricularByNome(componente.getId(),
 				componente.getNome().toUpperCase()) != null) {
-			result.rejectValue("nome", "Repeat.componente.nome",
-					"Já existe um componente curricular com esse nome");
+			result.rejectValue("nome", "Repeat.componente.nome", "Já existe um componente curricular com esse nome");
+			errors = true;
+		}
+
+		if (analiseComponente(componente, result, redirectAttributes, errors)) {
 			return "componente/editar";
 		}
 
 		componente.setNome(componente.getNome().toUpperCase());
 		componenteCurricularService.update(componente);
-		redirectAttributes.addFlashAttribute("info",
-				"Componente curricular atualizado com sucesso.");
+		redirectAttributes.addFlashAttribute("info", "Componente curricular atualizado com sucesso.");
 		return "redirect:/componente/listar";
 
 	}
 
 	@RequestMapping(value = "/{id}/excluir", method = RequestMethod.GET)
-	public String excluir(@PathVariable("id") Integer id,
-			RedirectAttributes redirectAttributes) {
+	public String excluir(@PathVariable("id") Integer id, RedirectAttributes redirectAttributes) {
 		ComponenteCurricular componente = componenteCurricularService.find(ComponenteCurricular.class, id);
-		if (componente != null) {
-			this.componenteCurricularService.delete(componente);
-			redirectAttributes.addFlashAttribute("info",
-					"Componente curricular removido com sucesso.");
+
+		if (integracaoCurricularService.getIntegracaoByIdComponente(id) == null) {
+			if (componente != null) {
+				this.componenteCurricularService.delete(componente);
+				redirectAttributes.addFlashAttribute("info", "Componente curricular removido com sucesso.");
+			}
+		} else {
+			redirectAttributes.addFlashAttribute("error", "O Componente Curricular (" + componente.getNome()
+					+ ") faz parte de Estruturas Curriculares de alguns cursos. Por favor, resolva estas dependências para tentar excluir este componente curricular.");
 		}
+
 		return "redirect:/componente/listar";
 	}
 
@@ -118,10 +138,9 @@ public class ComponenteCurricularController {
 		if (result.hasErrors()) {
 			return "componente/adicionar";
 		}
-		
-		if (componente.getCodigo().trim().isEmpty()) {
-			result.rejectValue("codigo", "Repeat.componente.codigo",
-					"Campo obrigatório.");
+
+		if (componenteCurricularService.getComponenteCurricularByNome(componente.getNome().toUpperCase()) != null) {
+			result.rejectValue("nome", "Repeat.componente.nome", "Já existe um componente curricular com esse nome");
 			errors = true;
 		}
 
@@ -130,38 +149,14 @@ public class ComponenteCurricularController {
 					"Já existe um componente curricular com esse código");
 			errors = true;
 		}
-		if (componente.getNome().trim().isEmpty()) {
-			result.rejectValue("nome", "Repeat.componente.nome",
-					"Campo obrigatório.");
-			errors = true;
-		}
-		if (componenteCurricularService.getComponenteCurricularByNome(componente.getNome()
-				.toUpperCase()) != null) {
-			result.rejectValue("nome", "Repeat.componente.nome",
-					"Já existe um componente curricular com esse nome");
-			errors = true;
-		}
-		
-		if (componente.getChPratica() == null) {
-			result.rejectValue("chPratica", "Repeat.componente.chPratica",
-					"Campo obrigatório.");
-			errors = true;
-		}
-		
-		if (componente.getChTeorica() == null) {
-			result.rejectValue("chTeorica", "Repeat.componente.chTeorica",
-					"Campo obrigatório.");
-			errors = true;
-		}
-		
-		if(errors == true){
+
+		if (analiseComponente(componente, result, redirectAttributes, errors)) {
 			return "componente/adicionar";
 		}
 
 		componente.setNome(componente.getNome().toUpperCase());
 		componenteCurricularService.save(componente);
-		redirectAttributes.addFlashAttribute("info",
-				"Componente curricular adicionado com sucesso.");
+		redirectAttributes.addFlashAttribute("info", "Componente curricular adicionado com sucesso.");
 		return "redirect:/componente/listar";
 	}
 
@@ -169,7 +164,7 @@ public class ComponenteCurricularController {
 	public String vincular(@PathVariable("id") Integer id, ModelMap modelMap) {
 		
 		ComponenteCurricular componente = this.componenteCurricularService.find(ComponenteCurricular.class, id);
-		
+
 		if (componente == null) {
 			return "redirect:/componente/listar";
 
@@ -211,39 +206,40 @@ public class ComponenteCurricularController {
 		}
 		return "/componente/listar";
 	}
-	
+
 	@RequestMapping(value = "/{id}/visualizar", method = RequestMethod.GET)
 	public String visualizar(@PathVariable("id") Integer id, ModelMap modelMap) {
 		
 		ComponenteCurricular componente = this.componenteCurricularService.find(ComponenteCurricular.class, id);
-		
+
 		if (componente == null)
 			return "redirect:/componente/listar";
+
 		
 		List<Titulo> basica = componente.getTitulosBibliografiasBasicas();
 		List<Titulo> complementar = componente.getTitulosBibliografiasComplementares();
-		
+
 		List<Bibliografia> bibliografias = componente.getBibliografias();
-		
+
 		List<IntegracaoCurricular> curriculos = componente.getCurriculos();
-		
+
 		HashMap<String, List<MetaCalculada>> metasCalculadasPorTitulo = new HashMap<String, List<MetaCalculada>>();
-		
+
 		List<ResultadoCalculo> resultados = calculoService.gerarCalculo();
-		
+
 		for (Bibliografia b : bibliografias) {
 			metasCalculadasPorTitulo.put(b.getTitulo().getNome(), new ArrayList<MetaCalculada>());
-			for(ResultadoCalculo resultadoCalculo : resultados){
-				if(resultadoCalculo.getTitulo().getId().equals(b.getTitulo().getId())){
-					for(MetaCalculada metaCalculada : resultadoCalculo.getMetasCalculadas()){
-						for(DetalheMetaCalculada detalheMetaCalculada: metaCalculada.getDetalhePar()){
-							if(detalheMetaCalculada.getComponenteCurricular().equals(componente.getNome())){
+			for (ResultadoCalculo resultadoCalculo : resultados) {
+				if (resultadoCalculo.getTitulo().getId().equals(b.getTitulo().getId())) {
+					for (MetaCalculada metaCalculada : resultadoCalculo.getMetasCalculadas()) {
+						for (DetalheMetaCalculada detalheMetaCalculada : metaCalculada.getDetalhePar()) {
+							if (detalheMetaCalculada.getComponenteCurricular().equals(componente.getNome())) {
 								metasCalculadasPorTitulo.get(resultadoCalculo.getTitulo().getNome()).add(metaCalculada);
 								break;
 							}
 						}
-						for(DetalheMetaCalculada detalheMetaCalculada: metaCalculada.getDetalheImpar()){
-							if(detalheMetaCalculada.getComponenteCurricular().equals(componente.getNome())){
+						for (DetalheMetaCalculada detalheMetaCalculada : metaCalculada.getDetalheImpar()) {
+							if (detalheMetaCalculada.getComponenteCurricular().equals(componente.getNome())) {
 								metasCalculadasPorTitulo.get(resultadoCalculo.getTitulo().getNome()).add(metaCalculada);
 								break;
 							}
@@ -252,47 +248,49 @@ public class ComponenteCurricularController {
 				}
 			}
 		}
-		
+
 		modelMap.addAttribute("bibliografia_basica", basica);
 		modelMap.addAttribute("bibliografia_complementar", complementar);
 		modelMap.addAttribute("metasCalculadas", metasCalculadasPorTitulo);
 		modelMap.addAttribute("curriculos", curriculos);
 		modelMap.addAttribute("componente", componente);
-		
+
 		return "componente/visualizar";
 	}
 
+	public List<Bibliografia> atualizarOuCriarBibliografia(String[] listaIdTitulo,
+			List<Bibliografia> bibliografiasAseremModificadas, ComponenteCurricular componente,
+			String tipoBibliografia) {
 
-	public List<Bibliografia> atualizarOuCriarBibliografia (String[] listaIdTitulo, List<Bibliografia> bibliografiasAseremModificadas, ComponenteCurricular componente, String tipoBibliografia){
 		int id_titulo;
-		
-		if(!listaIdTitulo[0].isEmpty()){
-		for (int i = 0; i < listaIdTitulo.length; i++) {
-			id_titulo = Integer.parseInt(listaIdTitulo[i]);
-			for (int j = 0; j < bibliografiasAseremModificadas.size(); j++) {
-				if (bibliografiasAseremModificadas.get(j).getTitulo().getId() == id_titulo) {
-					if (!bibliografiasAseremModificadas.get(j).getTipoBibliografia().equals(tipoBibliografia)) {
-						bibliografiasAseremModificadas.get(j).setTipoBibliografia(tipoBibliografia);
-						bibliografiaService.update(bibliografiasAseremModificadas.get(j));
-						bibliografiasAseremModificadas.remove(bibliografiasAseremModificadas.get(j));
-						listaIdTitulo[i]=null;
-						j=bibliografiasAseremModificadas.size()+1;
-						
-					}else{
-						bibliografiasAseremModificadas.remove(bibliografiasAseremModificadas.get(j));
-						listaIdTitulo[i]=null;
+
+		if (!listaIdTitulo[0].isEmpty()) {
+			for (int i = 0; i < listaIdTitulo.length; i++) {
+				id_titulo = Integer.parseInt(listaIdTitulo[i]);
+				for (int j = 0; j < bibliografiasAseremModificadas.size(); j++) {
+					if (bibliografiasAseremModificadas.get(j).getTitulo().getId() == id_titulo) {
+						if (!bibliografiasAseremModificadas.get(j).getTipoBibliografia().equals(tipoBibliografia)) {
+							bibliografiasAseremModificadas.get(j).setTipoBibliografia(tipoBibliografia);
+							bibliografiaService.update(bibliografiasAseremModificadas.get(j));
+							bibliografiasAseremModificadas.remove(bibliografiasAseremModificadas.get(j));
+							listaIdTitulo[i] = null;
+							j = bibliografiasAseremModificadas.size() + 1;
+
+						} else {
+							bibliografiasAseremModificadas.remove(bibliografiasAseremModificadas.get(j));
+							listaIdTitulo[i] = null;
+						}
+
 					}
-					
-				} 
+				}
+				if (listaIdTitulo[i] != null) {
+					Bibliografia biblio = new Bibliografia();
+					biblio.setComponenteCurricular(componente);
+					biblio.setTitulo(tituloService.find(Titulo.class, id_titulo));
+					biblio.setTipoBibliografia(tipoBibliografia);
+					bibliografiaService.save(biblio);
+				}
 			}
-			if (listaIdTitulo[i]!=null){
-				Bibliografia biblio = new Bibliografia();
-				biblio.setComponenteCurricular(componente);
-				biblio.setTitulo(tituloService.find(Titulo.class, id_titulo));
-				biblio.setTipoBibliografia(tipoBibliografia);
-				bibliografiaService.save(biblio);
-			} 
-		}
 		}
 		return bibliografiasAseremModificadas;
 	}
@@ -305,6 +303,7 @@ public class ComponenteCurricularController {
 		
 		if (componenteOrigem == null) {
 			return "redirect:/disciplina/listar";
+
 		}
 		List<Titulo> basica = componenteOrigem.getTitulosBibliografiasBasicas();
 		List<Titulo> complementar = componenteOrigem.getTitulosBibliografiasComplementares();
@@ -396,6 +395,46 @@ public class ComponenteCurricularController {
 				bibliografiaService.save(biblio);
 				}
 			}
+	}
+
+	private boolean analiseComponente(ComponenteCurricular componente, BindingResult result,
+			RedirectAttributes redirectAttributes, boolean errors) {
+
+		if (componente.getCodigo().trim().isEmpty()) {
+			result.rejectValue("codigo", "Repeat.componente.codigo", "Campo obrigatório.");
+			errors = true;
+		}
+
+		if (componente.getCodigo().length() < 5 || componente.getCodigo().length() > 12) {
+			result.rejectValue("codigo", "Repeat.componente.codigo",
+					"O código do Componente Curricular deve ter entre 5 e 12 caracteres");
+			errors = true;
+		}
+
+		if (componente.getChPratica() == null) {
+			result.rejectValue("chPratica", "Repeat.componente.chPratica", "Campo obrigatório.");
+			errors = true;
+		}
+
+		if (componente.getChTeorica() == null) {
+			result.rejectValue("chTeorica", "Repeat.componente.chTeorica", "Campo obrigatório.");
+			errors = true;
+		}
+
+		Pattern estruturaCodigo = Pattern.compile("[^A-Za-z0-9 ]");
+		Matcher matcher = estruturaCodigo.matcher(componente.getCodigo());
+		if (matcher.find()) {
+			result.rejectValue("codigo", "Repeat.componente.codigo",
+					"O campo código do Componente Curricular não pode possuir caracteres especiais.");
+			errors = true;
+		}
+
+		if (componente.getNome().trim().isEmpty()) {
+			result.rejectValue("nome", "Repeat.componente.nome", "Campo obrigatório.");
+			errors = true;
+		}
+
+		return errors;
 	}
 
 }
